Time is almost done Bug has occured when putting in time manually
